Be more careful about where to insert constraint assertions
If an assertion is required in a block that might not have precise flow
typing, e.g. inside an `if IsFeatureImplemented(...) ...`, then insert
the constraint into that block rather than attaching it to the val-spec
of the function, even if the constraint only contains top-level type
variables.

Also don't automatically add constraints to val-specs if we have
manually patched them.